+++ README.md
@@ -22,6 +22,6 @@ gulp test
 This runs the tests under tests directory.
 
 ### How to write a new validation rule
-1. Add a typescript file under ```azure-openapi-linter``` directory named same as the name of the rule. Add the ```id```, ```name```, ```severity```, ```category```,  ```mergeState```,  ```openapiType```,  ```appliesTo_JsonQuery``` properties to the rule. ```appliesTo_JsonQuery``` is the node(s) to which the rule needs to be applied. This is evaluated using JsonPaths. Please refer [here](https://www.npmjs.com/package/jsonpath#jsonpath-syntax) for a brief tutorial about JsonPaths.
-2. Next, implement the ```run``` method under the rule that actually does the validation. Add a reference to this script file under ```src\azure-openapi-linter\index.ts```.
-3. Lastly add a test case for the validation rule, by adding a test json/yaml under ```src\azure-openapi-linter\tests\resources``` and a script under ```src\azure-openapi-linter\tests``` depending on the type of the validation rule.
+1. Add a typescript file under ```azure-openapi-validator``` directory named same as the name of the rule. Add the ```id```, ```name```, ```severity```, ```category```,  ```mergeState```,  ```openapiType```,  ```appliesTo_JsonQuery``` properties to the rule. ```appliesTo_JsonQuery``` is the node(s) to which the rule needs to be applied. This is evaluated using JsonPaths. Please refer [here](https://www.npmjs.com/package/jsonpath#jsonpath-syntax) for a brief tutorial about JsonPaths.
+2. Next, implement the ```run``` method under the rule that actually does the validation. Add a reference to this script file under ```src\azure-openapi-validator\index.ts```.
+3. Lastly add a test case for the validation rule, by adding a test json/yaml under ```src\azure-openapi-validator\tests\resources``` and a script under ```src\azure-openapi-validator\tests``` depending on the type of the validation rule.
