+++ src/azure-openapi-validator/rule.ts
