+++ src/azure-openapi-validator/rules/ControlCharactersAreNotAllowed.ts
