+++ src/azure-openapi-validator/rules/DescriptionMustNotBeNodeName.ts
