+++ src/azure-openapi-validator/tests/individual-azure-tests.ts
@@ -6,7 +6,7 @@ import { Message } from '../../jsonrpc/types';
 import { suite, test, slow, timeout, skip, only } from "mocha-typescript";
 import { safeLoad } from "js-yaml";
 import { AutoRestPluginHost } from "../../jsonrpc/plugin-host";
-import { run } from "../../azure-openapi-linter";
+import { run } from "../../azure-openapi-validator";
 import {
   assertValidationRuleCount,
   collectTestMessagesFromValidator
