+++ src/azure-openapi-validator/tests/resources/ContainsControlCharacters.json
