+++ src/azure-openapi-validator/tests/resources/DescriptionSameAsPropertyName.json
