+++ src/index.ts
@@ -7,12 +7,12 @@ require("../utilities/polyfill.min.js");
 
 import { safeLoad } from "js-yaml";
 import { AutoRestPluginHost } from "./jsonrpc/plugin-host";
-import { run } from "./azure-openapi-linter";
-import { OpenApiTypes, MergeStates } from './azure-openapi-linter/rule';
+import { run } from "./azure-openapi-validator";
+import { OpenApiTypes, MergeStates } from './azure-openapi-validator/rule';
 
 async function main() {
   const pluginHost = new AutoRestPluginHost();
-  pluginHost.Add("azure-openapi-linter", async initiator => {
+  pluginHost.Add("azure-openapi-validator", async initiator => {
     const files = await initiator.ListInputs();
     const mergeState: string = await initiator.GetValue('merge-state');
     const openapiType: string = await initiator.GetValue('openapi-type');
